+++ commands/cmd_createversionstoreindex.go
@@ -13,6 +13,7 @@ import (
 func createVersionStoreIndex(
 	numWorkerCount int,
 	blobStoreURI string,
+	s3EndpointResolverURI string,
 	sourceFilePath string,
 	versionLocalStoreIndexPath string) ([]longtailutils.StoreStat, []longtailutils.TimeStat, error) {
 	const fname = "createVersionStoreIndex"
@@ -20,6 +21,7 @@ func createVersionStoreIndex(
 		"fname":                      fname,
 		"numWorkerCount":             numWorkerCount,
 		"blobStoreURI":               blobStoreURI,
+		"s3EndpointResolverURI":      s3EndpointResolverURI,
 		"sourceFilePath":             sourceFilePath,
 		"versionLocalStoreIndexPath": versionLocalStoreIndexPath,
 	})
@@ -33,7 +35,7 @@ func createVersionStoreIndex(
 	jobs := longtaillib.CreateBikeshedJobAPI(uint32(numWorkerCount), 0)
 	defer jobs.Dispose()
 
-	indexStore, err := remotestore.CreateBlockStoreForURI(blobStoreURI, "", jobs, numWorkerCount, 8388608, 1024, remotestore.ReadOnly, false)
+	indexStore, err := remotestore.CreateBlockStoreForURI(blobStoreURI, "", jobs, numWorkerCount, 8388608, 1024, remotestore.ReadOnly, false, longtailutils.WithS3EndpointResolverURI(s3EndpointResolverURI))
 	if err != nil {
 		return storeStats, timeStats, errors.Wrap(err, fname)
 	}
@@ -43,7 +45,7 @@ func createVersionStoreIndex(
 	timeStats = append(timeStats, longtailutils.TimeStat{"Setup", setupTime})
 
 	readSourceStartTime := time.Now()
-	vbuffer, err := longtailutils.ReadFromURI(sourceFilePath)
+	vbuffer, err := longtailutils.ReadFromURI(sourceFilePath, longtailutils.WithS3EndpointResolverURI(s3EndpointResolverURI))
 	if err != nil {
 		return storeStats, timeStats, errors.Wrap(err, fname)
 	}
@@ -73,7 +75,7 @@ func createVersionStoreIndex(
 		err = errors.Wrapf(err, "Cant serialize store index for `%s`", sourceFilePath)
 		return storeStats, timeStats, errors.Wrap(err, fname)
 	}
-	err = longtailutils.WriteToURI(versionLocalStoreIndexPath, versionLocalStoreIndexBuffer)
+	err = longtailutils.WriteToURI(versionLocalStoreIndexPath, versionLocalStoreIndexBuffer, longtailutils.WithS3EndpointResolverURI(s3EndpointResolverURI))
 	if err != nil {
 		return storeStats, timeStats, errors.Wrap(err, fname)
 	}
@@ -85,6 +87,7 @@ func createVersionStoreIndex(
 
 type CreateVersionStoreIndexCmd struct {
 	StorageURIOption
+	S3EndpointResolverURLOption
 	SourceUriOption
 	VersionLocalStoreIndexPathOption
 }
@@ -93,6 +96,7 @@ func (r *CreateVersionStoreIndexCmd) Run(ctx *Context) error {
 	storeStats, timeStats, err := createVersionStoreIndex(
 		ctx.NumWorkerCount,
 		r.StorageURI,
+		r.S3EndpointResolverURL,
 		r.SourcePath,
 		r.VersionLocalStoreIndexPath)
 	ctx.StoreStats = append(ctx.StoreStats, storeStats...)