+++ commands/cmd_initremotestore.go
@@ -13,13 +13,15 @@ import (
 func initRemoteStore(
 	numWorkerCount int,
 	blobStoreURI string,
+	s3EndpointResolverURI string,
 	hashAlgorithm string) ([]longtailutils.StoreStat, []longtailutils.TimeStat, error) {
 	const fname = "initRemoteStore"
 	log := logrus.WithFields(logrus.Fields{
-		"fname":          fname,
-		"numWorkerCount": numWorkerCount,
-		"blobStoreURI":   blobStoreURI,
-		"hashAlgorithm":  hashAlgorithm,
+		"fname":                 fname,
+		"numWorkerCount":        numWorkerCount,
+		"blobStoreURI":          blobStoreURI,
+		"s3EndpointResolverURI": s3EndpointResolverURI,
+		"hashAlgorithm":         hashAlgorithm,
 	})
 	log.Debug(fname)
 
@@ -31,7 +33,7 @@ func initRemoteStore(
 	jobs := longtaillib.CreateBikeshedJobAPI(uint32(numWorkerCount), 0)
 	defer jobs.Dispose()
 
-	remoteIndexStore, err := remotestore.CreateBlockStoreForURI(blobStoreURI, "", jobs, numWorkerCount, 8388608, 1024, remotestore.Init, false)
+	remoteIndexStore, err := remotestore.CreateBlockStoreForURI(blobStoreURI, "", jobs, numWorkerCount, 8388608, 1024, remotestore.Init, false, longtailutils.WithS3EndpointResolverURI(s3EndpointResolverURI))
 	if err != nil {
 		return storeStats, timeStats, errors.Wrap(err, fname)
 	}
@@ -69,6 +71,7 @@ func initRemoteStore(
 
 type InitRemoteStoreCmd struct {
 	StorageURIOption
+	S3EndpointResolverURLOption
 	HashingOption
 }
 
@@ -76,6 +79,7 @@ func (r *InitRemoteStoreCmd) Run(ctx *Context) error {
 	storeStats, timeStats, err := initRemoteStore(
 		ctx.NumWorkerCount,
 		r.StorageURI,
+		r.S3EndpointResolverURL,
 		r.Hashing)
 	ctx.StoreStats = append(ctx.StoreStats, storeStats...)
 	ctx.TimeStats = append(ctx.TimeStats, timeStats...)